Adjust news query and change filename to mongoUserService
