+++ routes/services.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const openUvService = require('../services/openUvService');
 const airVisualService = require('../services/airVisualService');
 const epicService = require('../services/epicService');
-const mongoService = require('../services/mongoService');
+const mongoService = require('../services/mongoUserService');
 const { newsAPI, guardianAPI } = require('../services/newsService');
 const News = require('../models/news');
 const mongoose = require('mongoose');
@@ -35,7 +35,9 @@ router.get('/services/epic/lastPhoto', async (req, res, next) => {
 // GET News API
 router.get('/services/news', async (req, res, next) => {
 	try {
-		const result = await newsAPI.getNews('"greta%20thunberg"OR"climate%20change"OR"environment"&language=en');
+		const result = await newsAPI.getNews(
+			'"greta%20thunberg"OR"climate%20change"OR"environment"&language=en&pageSize=40'
+		);
 
 		const news = result.data.articles.map(article => {
 			return {
@@ -71,7 +73,7 @@ router.get('/services/news', async (req, res, next) => {
 router.get('/services/guardian', async (req, res, next) => {
 	try {
 		const result = await guardianAPI.getNews(
-			'&section=environment&q="climate%20change"&page-size=20&show-fields=headline,byline,thumbnail,bodyText,trailText'
+			'&section=environment&q="climate%20change"&page-size=40&show-fields=headline,byline,thumbnail,bodyText,trailText'
 		);
 
 		const guardianNews = result.data.response.results.map(article => {
